Address comments

+++ tests/test.h
@@ -859,9 +859,9 @@ rd_kafka_resp_err_t test_delete_all_test_topics(int timeout_ms);
 void test_mock_cluster_destroy(rd_kafka_mock_cluster_t *mcluster);
 rd_kafka_mock_cluster_t *test_mock_cluster_new(int broker_cnt,
                                                const char **bootstraps);
-int test_mock_wait_maching_requests(
+size_t test_mock_wait_maching_requests(
     rd_kafka_mock_cluster_t *mcluster,
-    int num,
+    size_t num,
     int confidence_interval_ms,
     rd_bool_t (*match)(rd_kafka_mock_request_t *request, void *opaque),
     void *opaque);
